requirements.txt
# Core LangChain packages
langchain==0.1.20
langchain-openai==0.1.8
langchain-anthropic==0.1.15
langchain-community==0.0.38
langchain-core==0.1.52

# LangGraph for agent orchestration
langgraph==0.0.55

# Environment and configuration
python-dotenv==1.0.0

# Jupyter for interactive notebooks
jupyter==1.0.0
jupyterlab==4.1.5
ipywidgets==8.1.2

# Web interface (optional)
streamlit==1.32.0

# Vector stores for RAG
chromadb==0.4.24
faiss-cpu==1.7.4

# Tools and utilities
duckduckgo-search==5.3.0
wikipedia==1.4.0
requests==2.31.0

# Data processing
pandas==2.2.1
numpy==1.26.4

# Development and testing
pytest==8.1.1
pytest-asyncio==0.23.6
black==24.3.0
flake8==7.0.0

# Optional: For advanced features
# tiktoken==0.6.0  # OpenAI tokenizer
# openai==1.14.3   # Direct OpenAI SDK if needed
# anthropic==0.21.3  # Direct Anthropic SDK if needed

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.15
